requirements.txt
pandas 
sqlalchemy
neo4j
neo4j-driver
python-dotenv